requirements.txt
selenium
pymongo
pysocks
requests
stem
scrapy
scrapy_fake_useragent 
python-slugify
